+++ tests/test_console.py
@@ -116,9 +116,9 @@ app start --port 8000
 class TestLoadRecipeColonSyntax(unittest.TestCase):
     def test_load_recipe_with_colon_repetition(self):
         content = (
-            """web app setup-app:
-    --project one --home first
-    --project two
+            """web app setup:
+    - one --home first
+    - two
 web:
  - static collect
  - server start-app --host 1 --port 2
@@ -133,8 +133,8 @@ web:
             os.remove(temp_name)
 
         expected = [
-            ['web', 'app', 'setup-app', '--project', 'one', '--home', 'first'],
-            ['web', 'app', 'setup-app', '--project', 'two'],
+            ['web', 'app', 'setup', 'one', '--home', 'first'],
+            ['web', 'app', 'setup', 'two'],
             ['web', 'static', 'collect'],
             ['web', 'server', 'start-app', '--host', '1', '--port', '2'],
         ]
@@ -142,9 +142,9 @@ web:
 
     def test_load_recipe_colon_without_indentation(self):
         content = (
-            """web app setup-app:
---project one
---project two
+            """web app setup:
+- one
+- two
 web:
 - static collect
 - server start-app --host 1 --port 2
@@ -159,8 +159,8 @@ web:
             os.remove(temp_name)
 
         expected = [
-            ['web', 'app', 'setup-app', '--project', 'one'],
-            ['web', 'app', 'setup-app', '--project', 'two'],
+            ['web', 'app', 'setup', 'one'],
+            ['web', 'app', 'setup', 'two'],
             ['web', 'static', 'collect'],
             ['web', 'server', 'start-app', '--host', '1', '--port', '2'],
         ]